Prevent warning of setState on unmounted spring
Fixes #72.

If `endValue` calls an owner handler which sets some states, renders,
and umounts the spring, we'd have an already scheduled
`requestAnimationFrame` that does `setState` and would therefore warn.
This fixes that with a few checks. But in general, don't put side-effect
in `endValue`. This fix might not be guaranteed in the future!